feat: Add main file and compiled executable
Include the main file 4-main.c and the compiled executable 4-print_rev in the project directory. The main file contains a test case to validate the functionality of the print_rev function implemented in commit 1 (file 4-print_rev.c). Compilation was successful, and running the executable produces the expected output.

Task: 4