feat: Add main file and compiled executable
Include the main file `7-main.c` and the compiled executable `7-puts_half` in the project directory. The main file contains a test case to validate the functionality of the `puts_half` function implemented in commit 1 (file 7-puts_half.c). Compilation was successful, and running the executable produces the expected output.

Task: 7