feat: Add main file and compiled executable
Include the main file `8-main.c` and the compiled executable `8-print_array` in the project directory. The main file contains a test case to validate the functionality of the `print_array` function implemented in commit 1 (file 8-print_array.c). Compilation was successful, and running the executable produces the expected output.

Task: 8