feat: Add main file and compiled executable
Include the main file `4-main.c` and the compiled executable `4-rev_array` in the project directory. The main file contains test cases to validate the functionality of the `reverse_array` function implemented in commit 1 (file 4-rev_array.c). Compilation was successful, and running the executable produces the expected output.

Task: 4