feat: Add main file and compiled executable
Include the main file `5-main.c` and the compiled executable `5-string_toupper` in the project directory. The main file contains a test case to validate the functionality of the `string_toupper` function implemented in commit 1 (file 5-string_toupper.c). Compilation was successful, and running the executable produces the expected output.

Task: 5