feat: Add main file and compiled executable
Include the main file `6-main.c` and the compiled executable `6-cap` in the project directory. The main file contains a test case to validate the functionality of the `cap_string` function implemented in commit 1 (file 6-cap_string.c). Compilation was successful, and running the executable produces the expected output.

Task: 6