feat: Add main file and compiled executable
Include the main file `101-main.c` and the compiled executable `101-print_numbers` in the project directory. The main file contains test cases to verify the functionality of the `print_number` function implemented in commit 1 (file 101-print_number.c). Compilation was successful, and running the executable produces the expected output.

Task: 101