[Task 1] Add the _memcpy() function to the 1-memcpy.c file.
The _memcpy() function copies n bytes from the memory area pointed to by src to the memory area pointed to by dest. It takes a pointer to the destination memory area, a pointer to the source memory area, and the number of bytes to be copied as arguments. The function returns a pointer to the destination memory area after the copying operation.

This function provides a custom implementation of the memcpy() function, commonly found in the standard library, to copy memory areas.

Task: Add _memcpy() function.

Prototype: char *_memcpy(char *dest, char *src, unsigned int n);