[Task 1] Update the main.c file to test the _memcpy() function and compile the test file.
In the main.c file, a simple_print_buffer() function is implemented to print the contents of the buffer in hexadecimal format. The main function is modified to test the _memcpy() function. Two buffers, buffer and buffer2, are initialized. Then, _memcpy() is used to copy 10 bytes from buffer2 to buffer starting from offset 50.

Additionally, the 1-memcpy.c file is compiled with main.c to generate the 1-memcpy executable.

Task: Update main.c to test _memcpy() and compile test file.