[Task 3] Implement the _strspn() function in the 3-strspn.c file.
The _strspn() function gets the length of a prefix substring in the string s that consists only of bytes from the string accept. It takes a pointer to the string s and a pointer to the string accept as arguments. The function returns the number of bytes in the initial segment of s that consist only of bytes from accept.

This function provides a custom implementation of the strspn() function, commonly found in the standard library, to get the length of a prefix substring.

Prototype: unsigned int _strspn(char *s, char *accept);