[Task 3] Test the _strspn() function in the 3-main.c file.
In the main.c file, a test scenario is implemented. Two strings, s and f, are initialized with "hello, world" and "oleh," respectively. _strspn() is used to get the length of the prefix substring in s that consists only of bytes from f. The result is printed.

Additionally, the 3-strspn.c file is compiled with main.c to generate the 3-strspn executable.