[Task 4] Implement the _strpbrk() function in the 4-strpbrk.c file.
The _strpbrk() function searches the string s for any of the bytes in the string accept. It takes a pointer to the string s and a pointer to the string accept as arguments. The function returns a pointer to the first byte in s that matches one of the bytes in accept or NULL if no such byte is found.

This function provides a custom implementation of the strpbrk() function, commonly found in the standard library, to search a string for any of a set of bytes.

Prototype: char *_strpbrk(char *s, char *accept);