[Task 4] Test the _strpbrk() function in the 4-main.c file.
In the main.c file, a test scenario is implemented. Two strings, s and f, are initialized with "hello, world" and "world," respectively. _strpbrk() is used to search the string s for any of the bytes in f. The pointer to the matching byte in s is printed.

Additionally, the 4-strpbrk.c file is compiled with main.c to generate the 4-strpbrk executable.