[Task 5] Test the _strstr() function in the 5-main.c file.
In the main.c file, a test scenario is implemented. Two strings, s and f, are initialized with "hello, world" and "world," respectively. _strstr() is used to find the first occurrence of the substring f in the string s. The pointer to the beginning of the located substring is printed.

Additionally, the 5-strstr.c file is compiled with main.c to generate the 5-strstr executable.