[Task 100] Implement the set_string() function in the 100-set_string.c file.
The set_string() function sets the value of a pointer to a char. It takes a double pointer to a char and a pointer to a char as arguments. The function updates the value of the first pointer (char **) to point to the same memory location as the second pointer (char *). This way, both pointers will now point to the same string.

This function provides a custom implementation to set the value of a pointer to a char.

Prototype: void set_string(char **s, char *to);