[Task 100] Test the set_string() function in the 100-main.c file.
In the main.c file, two char pointers (s0 and s1) are initialized with strings "Bob Dylan" and "Robert Allen," respectively. The initial values of s0 and s1 are printed, and then the set_string() function is called to update the value of s1 to point to the same string as s0. The updated values of s0 and s1 are then printed.

Additionally, the 100-set_string.c file is compiled with main.c to generate the 100-set_string executable.