[Task 0] Implement the _puts_recursion() function
The _puts_recursion() function prints a string, followed by a new line, using recursion. It takes a pointer to a char (string) as an argument. The function prints each character of the string using _putchar() and then calls itself recursively with the remaining portion of the string until the null terminator is reached.

This function provides a custom implementation to print a string followed by a new line using recursion.

Prototype: void _puts_recursion(char *s);