[Task 1] Implement the _print_rev_recursion()
The _print_rev_recursion() function prints a string in reverse using recursion. It takes a pointer to a char (string) as an argument. The function first calls itself recursively with the string's next character until it reaches the null terminator. Then, it starts printing characters in reverse order as the recursive calls return.

This function provides a custom implementation to print a string in reverse using recursion.

Prototype: void _print_rev_recursion(char *s);