[Task 1] Test the _print_rev_recursion()
In the main.c file, the _print_rev_recursion() function is called with the string "\nColton Walker" to demonstrate its functionality. The provided string is printed in reverse order, followed by a new line.

Additionally, the 1-print_rev_recursion.c file is compiled with main.c and the _putchar.c file (which contains the _putchar() function) to generate the 1-print_rev_recursion executable.