[Task 4] Implement the _pow_recursion()
The _pow_recursion() function calculates and returns the value of x raised to the power of y using recursion. If y is less than 0, the function returns -1 to indicate an error.

This function provides a custom implementation to calculate the power of a number using recursion.

Prototype: int _pow_recursion(int x, int y);