[Task 100] Test the is_palindrome()
In the main.c file, the is_palindrome() function is called with various input strings ("level", "redder", "test", "step on no pets") to demonstrate its functionality. The results of the palindrome tests are printed.

Additionally, the 100-is_palindrome.c file is compiled with main.c to generate the 100-palindrome executable.