Added test program, compiled with static library
Included the 'main.c' test program file to validate the functionality of the utility functions implemented in the static library 'libmy.a'. The test program demonstrates practical usage scenarios for these functions.

Compilation of the test program was performed using the command: gcc -std=gnu89 main.c -L. -lmy -o quote

Upon successful compilation and linkage of the static library, the executable 'quote' was generated. Executing the program produced the expected result:
"At the end of the day, my goal was to be the best hacker"
    - Kevin Mitnick

This confirms the proper implementation of the utility functions within the static library.