Added script to create static library
Implemented the 'create_static_lib.sh' script, which automates the process of generating a static library named 'liball.a'. This script compiles all the .c files present in the current directory and creates a unified static library.

Upon executing the script, the 'liball.a' static library was successfully generated from the following .c files:
- 0-isupper.c
- 0-memset.c
- 0-strcat.c
- 100-atoi.c
- 1-isdigit.c
- 1-memcpy.c
- 1-strncat.c
- 2-strchr.c
- 2-strlen.c
- 2-strncpy.c
- 3-islower.c
- 3-puts.c
- 3-strcmp.c
- 3-strspn.c
- 4-isalpha.c
- 4-strpbrk.c
- 5-strstr.c
- 6-abs.c
- 9-strcpy.c
- _putchar.c

The resulting 'liball.a' library contains the compiled object files and can be utilized for various projects.