[Task 1] Implement program to print num of args
Implement a program in the 1-args.c file that prints the number of arguments passed into it. The program uses the argc variable to access the count of arguments and prints it to the standard output.

The program is compiled using the `gcc -Wall -pedantic -Werror -Wextra -std=gnu89 1-args.c -o nargs` command.