[Task 2]Implement program to print * received args
Implement a program in the 2-args.c file that prints all the arguments it receives. The program uses the argv array to access the arguments and prints each argument on a separate line, ending with a new line.

The program is compiled using the `gcc -Wall -pedantic -Werror -Wextra -std=gnu89 2-args.c -o args` command.