Task 1: Implement _strdup Function
Define the _strdup function which takes a string as a parameter and returns a pointer to a newly allocated space in memory containing a copy of the input string. Memory is allocated using malloc, and can be freed using free. The function returns NULL if the input string is NULL or if memory allocation fails.

Filename: 1-strdup.c