Task 2: Implement str_concat Function
Implement the str_concat function, which concatenates two strings and returns a pointer to a newly allocated space in memory. The resulting string contains the contents of s1, followed by the contents of s2, and is null-terminated. If either s1 or s2 is NULL, treat it as an empty string. The function returns NULL on failure.

Filename: 2-str_concat.c