Task 100: Implement argstostr Function
Create the argstostr function that concatenates all the arguments of a program. The function takes the number of arguments (ac) and an array of argument strings (av) as parameters. It returns a pointer to a new string where each argument is followed by a newline character ('\n'). Returns NULL if ac == 0 or av == NULL.

Filename: 100-argstostr.c