Task 101: Implement strtow Function
Define the strtow function, which splits a given string into words and returns a pointer to an array of strings. Each element of the array contains a single word, null-terminated. The last element of the array is set to NULL. Words are separated by spaces. Returns NULL if str == NULL or str == "".

Filename: 101-strtow.c