Task 1: Implement string_nconcat Function
Define the string_nconcat function which concatenates two strings: s1 followed by the first n bytes of s2, and null-terminated. The function returns a pointer to a newly allocated space in memory. If n is greater or equal to the length of s2, the entire string s2 is used. If NULL is passed, treat it as an empty string. Returns NULL if the function fails.

Filename: 1-string_nconcat.c