Task 2: Implement _calloc Function
Define the _calloc function that allocates memory for an array of 'nmemb' elements, each of 'size' bytes. The allocated memory is set to zero. The function returns a pointer to the allocated memory. If nmemb or size is 0, or if malloc fails, the _calloc function returns NULL.

Filename: 2-calloc.c