[0x0D_1]: Create a header file with a macro
Create a header file named `1-pi.h` that defines a macro named `PI` as an abbreviation for the token `3.14159265359`. This macro will be used to represent the mathematical constant pi in the program.

Task Number: 1
Files Created/Modified:
- 1-pi.h

Explanation:
The `PI` macro is defined in the header file to represent the value `3.14159265359`. This macro can be used to simplify the representation of the mathematical constant pi in the program, making the code more readable and easier to maintain.

The main program (`1-main.c`) includes the header file and demonstrates the usage of the `PI` macro by calculating the area of a circle. The result is then printed to the standard output.

This commit ensures that the `PI` macro is properly defined and demonstrates its usage through a simple program.