[0x0D_3]: Create a function-like macro
Define a function-like macro named `ABS(x)` that computes the absolute value of a given number `x`. This macro will be used to simplify the calculation of absolute values in the program.

Task Number: 3
Files Created/Modified:
- 3-function_like_macro.h

Explanation:
The `ABS(x)` macro is defined in the header file to calculate the absolute value of the number `x`. It uses the ternary operator to determine whether `x` is negative or not. If `x` is negative, it returns `-x`, otherwise, it returns `x`.

The main program (`3-main.c`) includes the header file and demonstrates the usage of the `ABS(x)` macro by calculating the absolute values of two numbers. The results are then printed to the standard output.

This commit ensures that the `ABS(x)` macro is properly defined and demonstrates its usage through a simple program.