[0x0E_2]Implement 'print_dog' Function
Create a function 'print_dog' in the '2-print_dog.c' file that prints the details of a 'struct dog' variable. The function takes a pointer to a 'struct dog' as an argument:
- 'd': A pointer to a 'struct dog' variable whose details need to be printed.

The function should follow these guidelines when printing the dog's information:
- If 'd' is NULL, do not print anything.
- If 'name' is NULL, print "Name: (nil)".
- If 'name' is not NULL, print "Name: [dog's name]".
- Print "Age: [dog's age]".
- If 'owner' is NULL, print "Owner: (nil)".
- If 'owner' is not NULL, print "Owner: [owner's name]".

The '2-main.c' file contains a sample program that demonstrates the usage of the 'print_dog' function. It initializes a 'struct dog' variable named 'my_dog' with specific values and then prints its details using the function.

Task Number: #2
File Names: 2-print_dog.c, 2-main.c