[0x10_0] Implement sum_them_all Function
Create a function 'sum_them_all' in the file '0-sum_them_all.c' that calculates the sum of a variable number of integers. The function takes the total number of integers as its first argument (n) and the variable arguments are the integers to be summed. If n is 0, the function returns 0. The file is provided in the repository '0x10-variadic_functions'.

Task Number: #0
File Name: 0-sum_them_all.c