[0x10_1] Implement print_numbers Function
Create a function 'print_numbers' in the file '1-print_numbers.c' that prints a variable number of integers separated by the provided separator string. The function takes a variable number of arguments using the variadic function feature and prints the numbers along with the separator. If the separator is NULL, it doesn't print it. The function ends with a new line.

Task Number: #2
File Name: 1-print_numbers.c