[0x10_2] Implement print_strings Function
Create a function 'print_strings' in the file '2-print_strings.c' that prints a variable number of strings separated by the provided separator string. The function takes a variable number of arguments using the variadic function feature and prints the strings along with the separator. If the separator is NULL, it doesn't print it. If a string is NULL, it prints (nil) instead. The function ends with a new line.

Task Number: #3
File Name: 2-print_strings.c