[0x10_3] Implement print_all Function
Create a function 'print_all' in the file '3-print_all.c' that prints a variable number of arguments based on the provided format string. The format string contains characters 'c', 'i', 'f', and 's' corresponding to the types char, integer, float, and char*. The function processes each argument based on its format and prints it accordingly. If a string is NULL, it prints (nil) instead. The function ends with a new line.

Task Number: #4
File Name: 3-print_all.c